pyproject.toml
[tool.poetry]
name = "fal"
version = "0.1.0"
description = "fal allows you to run python scripts directly from your dbt project."
homepage = "https://github.com/fal-ai/fal"
documentation = "https://docs.fal.ai"
repository = "https://github.com/fal-ai/fal"
readme = "README.md"
authors = [ "Features & Labels <hello@fal.ai>" ]
packages = [
        { include = "fal", from = "src" },
        { include = "faldbt", from = "src" },
        { include = "_fal_testing", from = "tests" }
]

keywords = [ "dbt", "pandas" ]
classifiers = [
    "Development Status :: 4 - Beta"
]

[tool.poetry.dependencies]
python = ">=3.7.1, <3.11"
arrow = "^1.2.0"
pandas = "^1.3.4"
google-cloud-bigquery = "^2.28.1"
google-cloud-bigquery-storage = "^2.9.1"
pyarrow = "^6.0.1"
pydantic = "^1.8.2"
PyYAML = "^6.0"
agate-sql = "^0.5.8"
firebase-admin = "^5.1.0"
sqlalchemy-bigquery = "^1.2.2"
requests = "^2.27.1"
posthog = "^1.4.5"
snowflake-sqlalchemy = "^1.3.3"
sqlalchemy-redshift = "^0.8.9"
deprecation = "^2.1.0"
astor = "^0.8.1"
duckdb-engine = "^0.1.8"
platformdirs = "^2.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3"
behave = "^1.2.6"
mock = "^4.0.3"
pytest-mock = "^3.7.0"
matplotlib = "^3.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fal = "fal.cli:cli"
